# Store topology spread constraints in metadata with labels.Selector